🚀 feat(firebase): add support for mc-v4-staging environment The .firebaserc file has been updated to include a new environment called mc-v4-staging. This environment has a single hosting configuration for mycure-suite-staging. The firebase-hosting-pull-request.yml file has been updated to use the mc-v4-stg project ID instead of mc-v4-prod. This allows pull requests to be deployed to the staging environment instead of the production environment.
🆕 chore(firebase-production.yml): add GitHub Actions workflow for deploying to Firebase Hosting
This commit adds a new GitHub Actions workflow file that deploys the application to Firebase Hosting on merge. The workflow is triggered by a push event with a tag that starts with 'v'. The workflow installs dependencies, creates an environment file, builds the application, installs Firebase tools, and deploys the application to Firebase Hosting. The deployment is done to the production environment using the mc-v4-prod project and the hosting target named 'production'.

🆕 chore(firebase-staging.yml): add Firebase Hosting deployment workflow
This commit adds a new workflow file that deploys the application to Firebase Hosting on merge to the main branch. The workflow installs dependencies, creates an environment file, builds the application, installs Firebase tools, and deploys the application to the staging environment.